+++ src/Crumb/index.js
@@ -5,5 +5,6 @@
  * Copyright 2014-2016, Tingle Team, Alinw.
  * All rights reserved.
  */
+import Crumb from './Crumb';
+export default Crumb;
 
-module.exports = require('./Crumb');
